CVS_SILENT more -Werror fixes
svn path=/trunk/kdenetwork/kopete/; revision=239250

+++ protocols/jabber/ui/dlgjabberservices.cpp
@@ -78,7 +78,7 @@ dlgJabberServices::dlgJabberServices (JabberAccount *account, QWidget *parent, c
 void dlgJabberServices::slotSetSelection (int row, int, int, const QPoint &)
 {
 
-	if(serviceTask && (row <= serviceTask->agents().count()))
+	if(serviceTask && (uint(row) <= serviceTask->agents().count()))
 	{
 		tblServices->clearSelection (true);
 		tblServices->addSelection (QTableSelection (row, 0, row, 1));
